Add Oikotie mapper
Add a mapper that maps the data received from Elasticsearch to
the django-oikotie `Apartment` dataclass so that an xml
representation can be created from the ES apartment data.

Optional values can be set as None.

Add tests for the mapper to check that the mapped values are of
correct type.

Refs ASU-289