Fix VAR_ROOT error when treated as a path
Using `environ.Path` with `default_var_root` seemed to cause local
Docker builds to fail with an error:

    AttributeError: 'str' object has no attribute '__root__'

The problem is `VAR_ROOT` is passed as an ENV string in `Dockerfile`,
while the Django settings expected it to be a path. The error
disappears if we don't convert the variable to a path.