ASU-773 Add authentication for the REST API
Creating a new profile now returns a masked profile ID
and a password, which can be used to request a JWT token.
The token is given in request headers in subsequent
requests to authenticate as the user connected to that
profile.

Viewing and editing profiles other than the user's own
is restricted.