ASU-832 Add pgcrypto extension and configuration
Adds the django-pgcrypto-fields library. The pgcrypto app is
also added to installed apps. This also enables the migration
from pgcrypto which automatically installs the pgcrypto extension
in PostgreSQL.

A public and private key should be defined in the .env file.
The .env.example file contains 1024-bit RSA keys that can be
used for development locally. These should not be used in
production.

For production environments, new keys should be generated
using the instructions given in README.md. Since they are
multi-line PEM strings, the newlines must be escaped with '\\n'.
See .env.example file for an example and README.md for a
snippet that converts the PEM to the correct format.