ASU-883 Encrypt application data
All application-related model fields (Application, Applicant,
ApplicationApartment) are encrypted. The migrations are done
without data loss by first creating new encrypted fields and
copying the existing data, then dropping the old ones.

For some Django field types django-pgcrypto-fields does not
have an equivalent. For UUIDs and booleans it's easy to create
new ones. But for positive small integer it's not possible,
because in the DB the field will be `bytea` and you cannot have
DB-level restrictions like "field >= 0" for such fields. So I
opted for the standard integer field which does not have such
DB-level restrictions.