ASU-919 Fix integrity error if identifier already exists
The get_or_create does not work the way the code was expecting it to
because an additional parameter was passed to the get_or_create
method. As a result, the code was trying to create a new row even
if the project or apartment ID was already found in the database.

One option would have been to use get_or_create first with just two
parameters, then update the record with the project or apartment. But
the same can be achieved by using the defaults parameter.