ASU-919: Make Profile primary key unencrypted
Having the primary key (UUID) of Profile as an encrypted field causes
the column to not be indexable, so queries with the primary key need to
perform a full table scan.

This is fixed by not encrypting the primary key. By having the primary
key unencrypted, the column can be indexed and lookups by the PK are
much faster.

Since the field is encrypted, we can't just alter the field in the
Django migrations. Instead we have to migrate through a temporary field.

The only model with Profile as foreign key is Application. Since the
values of the foreign key are already stored unencrypted, the
Application table does not require any changes.