ASU-1003: Fix applications getting cascade-deleted
Having `on_delete=models.CASCADE` for the profile foreign key caused
the applications to get deleted when a profile was deleted. We want
to keep all the applications even if the user's profile is deleted,
so it's better to use `on_delete=models.SET_NULL` instead and make
the field nullable.