ASU-1002: Add apartment queue services
Added service functions for adding applications to an apartment
queue. HASO application queues are automatically kept in order based on
the right of residence numbers. HITAS applications will be added to the
queue in the order they were added in.

Application can be submitted "late". Late applications are put at the
end of the queue and form their own "pool" of late applications which
are kept in order based on their right of residence numbers.

Applications can be removed from the queue. In that case the positions
of the applications behind the removed application will be incremented
by one.

Adding and removing can be done with a comment. In that case, a change
event will be recorded in the database.