ASU-1064: Add HITAS lottery logic
HITAS apartments are distributed via a lottery. For large apartments
(three rooms or more) priority is given to applications with children,
so shuffling is done in two pools. The first queue positions will go
to the apartments with children in random order, followed by the
remaining applicants in random order. For smaller apartments,
applications are shuffled in one pool.