Process zero price correctly in Etuovi
The test failed if the price were set to zero (0) value because the
mapper returned the None value instead of the zero value.

In the Etuovi integration document, price fields are required fields so
the price needs to be zero value instead of the None and generally, if
the price is zero, it should return the zero value always.