Add project installment templates to returned project data in API
Project installment templates' data is returned inside project data as
a list in "installment_templates" field.

The value of each instance is returned either in field "amount" (int,
in cents) or "percantage" (str) depending on the unit of the
installment template.

Currently the installment template data is included only in the project
detail endpoint, because the current easy solution doesn't probably
scale all that well to a large number of projects, and there is no
known use case which would need installment templates to be fetched for
multiple projects at the same time.