Add apartment_reservations root level API endpoint
Added API endpoint /v1/sales/apartment_reservations/<id>/ for fetching
a single reservation's data. The idea is that the endpoint is later
used when creating/updating/deleting reservations, but those aren't
implemented yet. There is also currently no use case for the list
endpoint, so that is not enabled either.