Add Customer update API
Implemented a basic API for updating an already existing Customer. It
is used with a PUT request to the Customer detail API endpoint.

The Profiles of the Customer are updated with the new data, and it is
currently not possible to switch the Profile(s) to other Profile
instances.

PATCH method is disabled, because a partial update would make the
needed logic more complicated, and there is no known use case for it
ATM.