Change apartment installment reference number behaviour in API a bit
Changed how ApartmentInstallment reference numbers behave when the
installments are created/updated via the API:

    * reference number is now read only
    * when an installment is created for the first time, a new reference
      number will be generated for it
    * when an installment is updated, it will keep its reference number
    * when an installment is deleted and then recreated, it will get a
      new reference number

Installment type is used to determine whether a POSTed installment is an
existing one or a new one, ie. if there is an existing installment with
the same type as a new one, those are considered to be the same.