Implement reservation state changes based on offer expiration
Implemented functionality that updates reservations' states based on
expiration status of their offers. This is done for reservations
everytime their order is updated. Also, there is a management
command for running the updation for all reservations, which is meant to
be run once a day as close as midnight (but not before) as possible.