Populate reservation right_of_residence from application when needed
ApartmentReservation.right_of_residence is now populated on creation
from the related application for reservations that are related to an
application.

For reservations that have an application this would not be
technically mandatory as the same value is in the application, but it
is nicer to have the value in the reservation then as well so we can
always just fetch it from there instead of needing to check whether an
application exists or not.