Fix reservation cancellation logic of other reservations
When a reservation is cancelled, the other reservations in the same
queue need to be processed and the new winner(s) need their states to be
updated. Before this fix, this logic for other reservations was
implemented only for reservations that have an application. Now when a
reservation is cancelled and there are no other reservations with an
application, the first reservation in the queue will be declared as the
current winner for both Hitas and Haso apartments.