Add valid_until to offer messages
Added valid_until date to the offer messages that can be generated for
reservations. The value can be given in the offer message API request
with query parameter valid_until=<YYYY-MM-DD>. If the parameter is not
used, the value of the reservation's offer instance will be used if one
exists.

This functionality is needed because the Sales UI needs to fetch offer
messages with valid until dates also for reservations that do not have
an existing offer yet.