Fix tests to use correct api client
There are two salesperson api clients that can be used in the tests, one
meant for testing API endpoints used by the Drupal site, and one for
testing API endpoints used by the Sales UI. This commit fixes the usage
of those two in the tests, so that the tests will become more thorough.
e # Please enter the commit message for your changes. Lines starting