Refactor installment update
Before this commit ProjectInstallmentTemplate and ApartmentInstallment
objects were recreated everytime they where updated using the API. It
was implemented this way back in the day because it was the easiest way
and made sense then.

Now that there are all kinds of new data related to installments, like
audit logs for example, recreating the objects on every update causes
increasing unnecessary complexity. To fix that, this commit changes
installment objects to be actually updated instead of recreated when
updated via the API.