Add new reservation cancellation reasons
Added new choices LOWER_PRIORITY and OFFER_REJECTED to reservation
cancel reasons. The former is used when a reservation is cancelled for
having a lower priority than a reserved reservation, and the latter when
the reservation has an offer that was rejected.

Both of these new values are used automatically, and Sales UI users are
not allowed to manually set them.