Add API endpoint for fetching apartment's all reservations
Added API endpoint /v1/sales/apartments/<apartment UUID>/reservations/
which returns the given apartment's all reservations sorted by
list_position. In practice this is meant to replace project detail
endpoint reservations field, which was removed earlier. This endpoint's
data does not have has_multiple_winning_reservations field, but other
than that the data is the same as in the precessor.