HASO PDF: Fix type of living_area and use real m^2 char
Fix the type of living_area to float, since that's how it's currently
used at least: The value to this field is filled from "living_area"
field of ApartmentDocument (ES doc), which defines the field as Float.

Also change the usage of "m\u002b" to "m²" to make it easier to read.
There should be no problem using UTF-8 coded unicode code points in the
Python code str values.

Refs ASU-1642