pytest settings: Add -ra flag
Add "-ra" as default flag for pytest runs to make sure that warnings are
visible in the test run summary output.

Also remove the pytest.ini and move the DJANGO_SETTINGS_MODULE setting
from it to the setup.cfg, which has other pytest settings, because
otherwise the other pytest settings in setup.cfg won't be effective.