requirements: Update pip-tools, add Makefile
Update the version of pip-tools to 6.13.0 so that it's compatible with
newer Pip.

Add "-c requirements.txt" to requirements-dev.in and
requirements-prod.in so that the generated txt files won't get package
versions that conflict with the versions in requirements.txt.

Add Makefile to allow updating all requirements*.txt files with a single
"make" call.