Fix isort config and resort imports
Fix the isort configuration is setup.cfg and run "isort ." to resort all
imports.

There was a problem with the previous configuration which caused
standard library modules to be classified as thirdparty modules and that
messed up the import grouping.  Fix the configuration by removing
overrides of the standard library modules in setup.cfg with the
"known_standard_library" setting.  Also remove the unneeded
"known_first_party" setting.