isort also migrations
Change isort settings in setup.cfg so that migrations are no longer
excluded and run "isort ." to resort imports in the migration files.

There's no point excluding migrations, especially since they are not
excluded by black either.

There was also "venv" specified to be excluded with the "skip" option,
but that is in the default value of the "skip" option -- and so is
".git" which wasn't excluded previously, because the "skip" option was
overridden.