audit_log: Fix LargeTablePaginator edge case
The LargeTablePaginator uses PostgreSQL's "reltuples" method for
estimating the number of rows in the table.  When the table in question
hasn't been analyzed or vacuumed, the returned count will be -1 (at
least on PostgreSQL >= 14).  This case was previously handled in the
test case so that the test expects the returned count to always be -1
for PostgreSQL >= 14.  This has two problems:

  1. The paginator count shouldn't ever return negative value

  2. If the test table was already analyzed or vacuumed, then the
     returned count is no longer -1 even on PostgreSQL 14 and
     the test will fail.

Fix this doing the -1 check in the paginator itself and fallback to
normal "SELECT COUNT(*)" method in that case.

The -1 case (and the related fallback to normal COUNT) happens during
test runs sometimes, and it could make a difference whether inserts have
been done to the AuditLog table by other tests.  It's possible to see
when this happens by using --log-cli-level=WARNING flag for pytest.

Also add another test case which also inserts some rows to the AuditLog
table before calling the paginator count.  And add order_by to the old
and the new test case so that Django won't emit a warning about using a
paginator on unordered queryset.

Note that the tests cannot assert the returned count to be exactly same
as the real count, since it is actually sometimes out of date.  When I
was running the tests, it was 0 on some test runs and the actual count
on others.  However, it shouldn't be -1 anymore, since that case is now
handled in the paginator with the fallback.