invoicing: Fix percentage_specifier translations in error message
When the translations have been compiled ("manage.py compilemessages"),
the error messages generated by the ProjectInstallmentTemplateSerializer
used to change, because they included the string value of the
"percentage_specifier" variable, which is an instance of the enum
InstallmentPercentageSpecifier and being an enumfields.Enum it's string
value is the translated label.  This caused the related two test cases
to fail.

Fix this issue by using the "name" of the enum instead of its label in
the error message.  Amend the test cases accordingly.