SaleReportAPIView: Fix date range timezones
The start and end date passed to the SaleReportAPIView are parsed to
naive date objects and the passed to Django ORM query as arguments to
timestamp__range.  That's a bit ambiguous and yields a warning, since
the timestamp field is a datetime field, which always has a timezone in
Django but the range endpoints are given without timezone.

Fix this by converting the start and end dates explicitly to the default
timezone as is configured in the settings (Europe/Helsinki).

This doesn't actually change the behaviour otherwise, but gets rid of
the warning from Django (and is easier to understand what timezones are
actually used in the code).