create_pdf: Fix non-annotated page case
The _set_pdf_fields (which is used in the create_pdf function) iterates
over all pages and then used to use page.Annots to access the
annotations on each page, however that fill fail if some of the pages
doesn't have any annotations, since the attribute doesn't even exists in
that case.  Fix this by using getattr.