asko_import: Disable invoice_number MIN value validation
Set the limit value of the MinValueValidator to 1 for the invoice_number
field of the ApartmentInstallment during the import.

This allows importing the ApartmentInstallments even though they have
their invoice_numbers being smaller than the 730000001 which is defined
as the smallest possible value in the field validators of the model.