asko_import: Allow progressive imports (--commit-each)
Add flags --commit-each and --skip-imported which make it possible to
save the progress so far for each of the successfully imported models
and continue from there with the next import.

This saves time in development, since now I don't have to run 45 minutes
of the successful models before finding error in one of the later
models.