asko_import: Initialize incrementing list_positions
When importing ApartmentReservation objects, make sure that the
list_position is initialized with unique value for each apartment_uuid
and reservation so that the reservations can be saved before list
positions are restored from AsKo import files.

Note that these numbers will be overridden in the later phase of the
import (see _set_reservation_positions function), but the code uses
positions which are multiplied by 10000 to make sure that they stick out
if any of them is left unupdated.