asko_import: Skip ApartmentInstallments with no apartment_reservation
In the AsKo data there is several ApartmentInstallment rows which don't
have a value in the "apartment_reservation" field.  Those cannot be
imported so we must skip them.

It's possible that we could somehow link these through some other field
in the future, if we get e.g. customer info, but at the moment the
import data doesn't have that data so skipping these rows is the only
possibility.