Add right_of_residence_is_old_batch field
The boolean field is added to all models that have `right_of_residence`
field, and it includes information whether the number is of the old
batch (granted on 31.8.2023 or before) or the new batch (granted
1.9.2023 or later). There is also property
`right_of_residence_ordering_number` which should be used in doing
ordering based on right of residence numbers. This makes sure the old
batch will be placed before the new batch.