asko_import: Handle right_of_residence_is_old_batch
Set the right_of_residence_is_old_batch to True for all applications,
customers and reservations with a right_of_residence number.

Note: The objects which don't have a right_of_residence number will get
a value of None to the is-old-batch field, since that is handled by the
overridden save method in the CommonApplicationData model mixin.