asko_import: Don't force queue pos 1 to RESERVED
It seems that there's a bunch of apartments which are in
FREE_FOR_RESERVATION state in AsKo, i.e. all reservations of such
apartment are either SUBMITTED or CANCELED.  It's better to import them
in the same state.  If that causes problems later, we can easily convert
the reservation on queue position 1 as RESERVED at that point.