Check requirements files in CI
Make sure that the requirements*.txt files are correctly compiled from
their *.in counterparts.

Add a script that checks the requirements files and hook it to CI in
GitHub actions and add it also to the Makefile.  (Note: The requirements
files are compiled with Python 3.8 so the check will fail with other
Python versions.)

This should make it safer to merge updates to requirements.txt, since
now we can be sure that the version change doesn't cause any conflicts
between the package versions.