Fix couple ElasticSearch searches without index
These searches were doing the search to all indices, which is
problematic if there are several indices in the used ES instance.

This was causing tests to fail for me locally, since I had one ES index
which had an apartment A0 without project_uuid and the code fails if
there is no project_uuid.