applications: Set customer ssn_suffix from applicant
As part of the application creation business logic, in the
`create_application` function, make sure that the NIN (aka HETU in
Finnish) is added to the user profile if it didn't have one yet.

It could be better to fill the NIN to the profile initially when
creating the Profile object through the "POST /v1/profiles" endpoint,
but that would need changes to the API.  If NIN is introduced as a
required field, then also the Drupal side would have to be updated.

This affects both of application creating endpoints:
 * POST /v1/applications/
 * POST /v1/sales/applications/