Fix slow project applicant validation
When a new application is created, the applicants are validated to make
sure the same person (same date of birth and SSN suffix) hasn't already
applied to the same project. This validation was painfully slow before,
it took over a minute in production, because Postgres ended up
decrypting all applicants in the database.

The new implementation fetches the project's all DOBs and SSN suffixes
first than checks those in Python to make sure only the project's
applicants are decrypted. This makes the validation a lot faster when
there is a lot of data in the db.