Fix adding application to queue with cancelled reservations
The previous fix on commit 72ad1913 wasn't enough, because the list
position will still need adjusting even though the reservation is
cancelled.  This is because (apartment_uuid, list_position) is a unique
key in the database and if we don't adjust the list positions there will
be a duplicate key.

Fix the issue by making the code to adjust both the queue positions and
the list positions.

Additionally, fix the _calculate_queue_position function so that it
won't read the queue positions of the cancelled reservations, because
they are None and mess up the calculations.

Also add a test case to make sure that this fix works better.