invoicing: Fix UTF-8 to Latin1 in SAP payment fetching
The payment data file seems to be encoded as Latin1 (aka ISO-8859-1)
instead of UTF-8.  Fix the fetcher to decode it correctly using the
"latin1" as the encoding name.

Also update the test and move the test data to a different file, adding
couple umlauts (ä letters) so that decoding it as UTF-8 would fail.
Additionally, the line endings are CRLF in the payment data file, so
update that too.