HASO PDF: Update HASO contract template and fields
Replace the PDF template file with a new one and update the field
mapping accordingly.

Field changes:
 * "approver": New field, hardcoded as "Helsingin kaupunki"
 * "occupant_1_signing_text": New field, filled with profile 1 name
 * "occupant_2_signing_text": New field, filled with profile 2 name
 * "signing_place_and_time": replaces "signing_place" and "signing_time"
 * "signing_text": Removed
 * "right_of_occupancy_payment_text": Removed (not present in new)

Also update flake8 configuration in setup.cfg to exclude the long-lines
error E501 for the test file, since the expected lines from the PDF are
easier to work with when formatted this way.