pdf.py: Preserve PDF as much as possible
The previous implementation removed e.g. the bookmarks (Table of
Contents) from the PDF file, because it just copied the page contents,
but not any metadata from the template to the new PDF.  Improve the PDF
creation so that if there is no multiple data entries repeated, then we
copy the template PDF as-is and fill the fields there.

When filling the same PDF template with multiple data entries, the code
does the same copying of only the pages to an empty PDF file, because
table of contents would have to be updated otherwise.  This is currently
only used when generating a PDF file for invoice installments.