docker-compose: Remove publicly open ports
Remove the expose of PostgreSQL port totally, since it shouldn't be
needed from the outside and it conflicts with a local PostgreSQL server,
if there is one running on the host machine.

Expose the ElasticSearch port 9200 and Django app port 8081 only to
localhost so that they are not accidentally open to public Internet on
the developer's machine.