Use project_contract_combined_terms in the PDFs
It seems that in ElasticSearch the data in the
project_contract_other_terms field is not enough for the PDF, but we
need to combine also data from two other fields.  Add the two other
fields to ApartmentDocument and a property attribute which combines
those fields with couple newlines.

Update the PDF generation code to use the combined field.