Allow disabling applicant duplicate validation
This is useful for testing, as it allows us to create multiple
applications for the same applicant.

This is not allowed in production, as it would allow an applicant to
submit multiple applications for the same apartment, but there was a bug
that caused multiple applications for same applicant to be created even
when the validation was enabled.  Therefore we now need this setting to
be able to test situations where there are multiple applications for the
same applicant.