Hitas PDF: Fix numbered payment percentages
Before this change numbered payment (PAYMENT_1-PAYMENT_7) percentage
values came straight from the corresponding project installment
templates' values. That was not ideal, because there is nothing
preventing editing prices for an apartment manually, which would have
resulted in wrong percentages being shown. Now the percentage values are
actually calculated for each row and rounded to 1 decimal place.

A possible flexible payment is an exception: it's value is

100% - (sum of all other numbered payments' percentages rounded to 1
decimal place)

So when a flexible payment is used, the payments' percentages sum always
matches precisely 100%. This functionality has been agreed with HKI.