Fix late application queue
Late applications queue in HASO projects were not
properly ordered.
Late submitted applications have their own queue logic, they
are placed under the normal applications.

Late applications are the ones added by the admin after the lottery
has been executed.

Change the logic so that the late applications will go below the non late ones.
They are ordered by right or residency number.
If there is an offered reservation, it will lock on top of the list,
even if a new appliction is added with a right of residency number smaller,
it will be placed under the locked reservation.

Added new field to reservations submitted_late to differentiate
late reservations.

Updated tests and added a validation for users with no right of residency number.