Fix parsing of line when fetching payment amount
Previously, the parser misread the payment value, appending an extra zero.
This was due to an incorrect parsing position within the string, which has
the format ie:  00062905000.
The last digit is not part of the actual amount. The first 10 are.

The parser is now adjusted to move two positions to the left, correctly
extracting the payment value, with the lenght of 10 characters.

Added documentation table for formatting the payment on Liner Parser class.