fix: decrypt application
Includes Applicant, Application and ApplicationApartment.

This one was fairly interesting because there are encrypted fields with
a default value. django-pgcrypto-fields does not support default values
(on the db level). See:
 https://github.com/incuna/django-pgcrypto-fields/issues/298

The way django uses `default` has a catch - if the
field is nullable, django does nothing on the DB side with DEFAULT. If
the field is NOT nullable, then django will apply the DEFAULT and
immediately after DROP DEFAULT. For an encrypted field this will not
work as it will attempt to set the primitive value to the encrypted
field which has a type of bytea.

One way to encounter the catch is to remove a field that is NOT NULL
and has a DEFAULT:
the reverse migration will attempt to create the field with the DEFAULT
value --> kaboom.

Attempting to ALTER field from NOT NULL + DEFAULT -> NULL will
also fail. Django attempts to use the previous default in this
migration.

So the correct way is to go in two steps and first remove the default
and then allow nulls:

NOT NULL + DEFAULT -> NOT NULL -> NULL